Added acceptFunOnArbitrumFilter
