+++ src/resolvers/projectResolver.ts
@@ -530,6 +530,12 @@ export class ProjectResolver {
           // Add this to make sure works on Staging
           networkIds.push(NETWORK_IDS.CELO_ALFAJORES);
           return;
+
+        case FilterField.AcceptFundOnArbitrum:
+          networkIds.push(NETWORK_IDS.ARBITRUM_MAINNET);
+          networkIds.push(NETWORK_IDS.ARBITRUM_SEPOLIA);
+          return;
+
         case FilterField.AcceptFundOnPolygon:
           networkIds.push(NETWORK_IDS.POLYGON);
           return;
